fix some gym envs

+++ examples/pybullet/gym/pybullet_envs/robot_bases.py
@@ -22,9 +22,9 @@ class XmlBasedRobot:
 		self.robot_body = None
 
 		high = np.ones([action_dim])
-		self.action_space = gym.spaces.Box(-high, high, dtype=np.float32)
+		self.action_space = gym.spaces.Box(-high, high)
 		high = np.inf * np.ones([obs_dim])
-		self.observation_space = gym.spaces.Box(-high, high, dtype=np.float32)
+		self.observation_space = gym.spaces.Box(-high, high)
 
 		#self.model_xml = model_xml
 		self.robot_name = robot_name
